fix: 초기화면 진입 막힌느 버그 수정
